moved last modifications into source (were wrongly inserted in compiled file)
