+++ smart-app-banner.js
@@ -1,321 +1,10 @@
-(function (f) {
-  if (typeof exports === "object" && typeof module !== "undefined") {
-    module.exports = f()
-  } else if (typeof define === "function" && define.amd) {
-    define([], f)
-  } else {
-    var g;
-    if (typeof window !== "undefined") {
-      g = window
-    } else if (typeof global !== "undefined") {
-      g = global
-    } else if (typeof self !== "undefined") {
-      g = self
-    } else {
-      g = this
-    }
-    g.SmartBanner = f()
-  }
-})(function () {
-  var define, module, exports;
-  return (function e(t, n, r) {
-    function s(o, u) {
-      if (!n[o]) {
-        if (!t[o]) {
-          var a = typeof require == "function" && require;
-          if (!u && a) return a(o, !0);
-          if (i) return i(o, !0);
-          var f = new Error("Cannot find module '" + o + "'");
-          throw f.code = "MODULE_NOT_FOUND", f
-        }
-        var l = n[o] = {
-          exports: {}
-        };
-        t[o][0].call(l.exports, function (e) {
-          var n = t[o][1][e];
-          return s(n ? n : e)
-        }, l, l.exports, e, t, n, r)
-      }
-      return n[o].exports
-    }
-    var i = typeof require == "function" && require;
-    for (var o = 0; o < r.length; o++) s(r[o]);
-    return s
-  })({
-    1: [function (require, module, exports) {
-      var extend = require('xtend/mutable');
-      var q = require('component-query');
-      var doc = require('get-doc');
-      var root = doc && doc.documentElement;
-      var cookie = require('cookie-cutter');
-
-
-      // platform dependent functionality
-      var mixins = {
-        ios: {
-          appMeta: 'apple-itunes-app-meta',
-          iconRels: ['apple-touch-icon-precomposed', 'apple-touch-icon'],
-          getStoreLink: function () {
-            return this.options.url.ios;
-          }
-        },
-        android: {
-          appMeta: 'google-play-app',
-          iconRels: ['android-touch-icon', 'apple-touch-icon-precomposed', 'apple-touch-icon'],
-          getStoreLink: function () {
-            return this.options.url.android;
-          }
-        }
-      };
-
-      var SmartBanner = function (options) {
-        var userAgent = navigator.userAgent;
-        this.options = extend({}, {
-          daysHidden: 15,
-          daysReminder: 90,
-          appStoreLanguage: 'us', // Language code for App Store
-          button: 'OPEN', // Text for the install button
-          store: {
-            ios: 'On the App Store',
-            android: 'In Google Play'
-          },
-          price: {
-            ios: 'FREE',
-            android: 'FREE'
-          },
-          force: false // put platform type (ios, android, etc.) here for emulation
-        }, options || {});
-
-        if (this.options.force) {
-          this.type = this.options.force;
-        } else if (userAgent.match(/iPad|iPhone|iPod/i) !== null) {
-          if (userAgent.match(/Safari/i) !== null ||
-            (userAgent.match(/CriOS/i) !== null ||
-              Number(userAgent.substr(userAgent.indexOf('OS ') + 3, 3)
-                .replace('_', '.')) < 6)) {
-            this.type = 'ios';
-          } // Check webview and native smart banner support (iOS 6+)
-        } else if (userAgent.match(/Android/i) !== null) {
-          this.type = 'android';
-        }
-
-        // Don't show banner if device isn't iOS or Android, website is loaded in app, user dismissed banner, or we have no app id in meta
-        if (!this.type || navigator.standalone || cookie.get('smartbanner-closed') || cookie.get('smartbanner-installed')) {
-          return;
-        }
-
-        extend(this, mixins[this.type]);
-
-        if (!this.parseAppId()) {
-          return;
-        }
-
-        this.create();
-        this.show();
-      };
-
-      SmartBanner.prototype = {
-        constructor: SmartBanner,
-
-        create: function () {
-          var link = this.getStoreLink();
-          var inStore = this.options.price[this.type] + ' - ' + this.options.store[this.type];
-          var icon;
-          if (this.options.icon) {
-            icon = this.options.icon;
-          } else {
-            for (var i = 0; i < this.iconRels.length; i++) {
-              var rel = q('link[rel="' + this.iconRels[i] + '"]');
-              if (rel) {
-                icon = rel.getAttribute('href');
-                break;
-              }
-            }
-          }
-
-          var sb = doc.createElement('div');
-          sb.className = 'smartbanner smartbanner-' + this.type;
-
-          sb.innerHTML = '<div class="smartbanner-container">' +
-            '<a href="javascript:void(0);" class="smartbanner-close">&times;</a>' +
-            '<span class="smartbanner-icon" style="background-image: url(' + icon + ')"></span>' +
-            '<div class="smartbanner-info">' +
-            '<div class="smartbanner-title">' + this.options.title + '</div>' +
-            '<div class="smartbanner-description">' + this.options.description + '</div>' +
-            '<div>' + this.options.author + '</div>' +
-            '<span>' + inStore + '</span>' +
-            '</div>' +
-            '<a href="' + link + '" class="smartbanner-button">' +
-            '<span class="smartbanner-button-text">' + this.options.button + '</span>' +
-            '</a>' +
-            '</div>';
-
-          //there isn’t neccessary a body
-          if (doc.querySelector('.body-wrapper')) {
-            doc.querySelector('.body-wrapper')
-              .appendChild(sb);
-            doc.body.classList.add('smart-banner');
-          } else if (doc) {
-            doc.addEventListener('DOMContentLoaded', function () {
-              doc.querySelector('.body-wrapper')
-                .appendChild(sb);
-              doc.body.classList.add('smart-banner');
-            });
-          }
-
-          q('.smartbanner-button', sb)
-            .addEventListener('click', this.install.bind(this), false);
-          q('.smartbanner-close', sb)
-            .addEventListener('click', this.close.bind(this), false);
-
-        },
-        hide: function () {
-          root.classList.remove('smartbanner-show');
-        },
-        show: function () {
-          root.classList.add('smartbanner-show');
-        },
-        close: function () {
-          this.hide();
-          doc.body.classList.remove('smart-banner');
-          if (this.options.daysHidden > 0) {
-            cookie.set('smartbanner-closed', 'true', {
-              path: '/',
-              expires: +new Date() + this.options.daysHidden * 1000 * 60 * 60 * 24
-            });
-          }
-        },
-        install: function () {
-          this.hide();
-          doc.body.classList.remove('smart-banner');
-          if (this.options.daysReminder > 0) {
-            cookie.set('smartbanner-installed', 'true', {
-              path: '/',
-              expires: +new Date() + this.options.daysReminder * 1000 * 60 * 60 * 24
-            });
-          }
-        },
-        parseAppId: function () {
-          var meta = q('meta[name="' + this.appMeta + '"]');
-          if (!meta) {
-            return;
-          }
-
-          if (this.type === 'windows') {
-            this.appId = meta.getAttribute('content');
-          } else {
-            this.appId = /app-id=([^\s,]+)/.exec(meta.getAttribute('content'))[1];
-          }
-
-          return this.appId;
-        }
-      };
-
-      module.exports = SmartBanner;
-
-    }, {
-      "component-query": 2,
-      "cookie-cutter": 3,
-      "get-doc": 4,
-      "xtend/mutable": 6
-    }],
-    2: [function (require, module, exports) {
-      function one(selector, el) {
-        return el.querySelector(selector);
-      }
-
-      exports = module.exports = function (selector, el) {
-        el = el || document;
-        return one(selector, el);
-      };
-
-      exports.all = function (selector, el) {
-        el = el || document;
-        return el.querySelectorAll(selector);
-      };
-
-      exports.engine = function (obj) {
-        if (!obj.one) throw new Error('.one callback required');
-        if (!obj.all) throw new Error('.all callback required');
-        one = obj.one;
-        exports.all = obj.all;
-        return exports;
-      };
-
-    }, {}],
-    3: [function (require, module, exports) {
-      var exports = module.exports = function (doc) {
-        if (!doc) doc = {};
-        if (typeof doc === 'string') doc = {
-          cookie: doc
-        };
-        if (doc.cookie === undefined) doc.cookie = '';
-
-        var self = {};
-        self.get = function (key) {
-          var splat = doc.cookie.split(/;\s*/);
-          for (var i = 0; i < splat.length; i++) {
-            var ps = splat[i].split('=');
-            var k = unescape(ps[0]);
-            if (k === key) return unescape(ps[1]);
-          }
-          return undefined;
-        };
-
-        self.set = function (key, value, opts) {
-          if (!opts) opts = {};
-          var s = escape(key) + '=' + escape(value);
-          if (opts.expires) s += '; expires=' + opts.expires;
-          if (opts.path) s += '; path=' + escape(opts.path);
-          doc.cookie = s;
-          return s;
-        };
-        return self;
-      };
-
-      if (typeof document !== 'undefined') {
-        var cookie = exports(document);
-        exports.get = cookie.get;
-        exports.set = cookie.set;
-      }
-
-    }, {}],
-    4: [function (require, module, exports) {
-      /**
-       * @module  get-doc
-       */
-
-      var hasDom = require('has-dom');
-
-      module.exports = hasDom() ? document : null;
-    }, {
-      "has-dom": 5
-    }],
-    5: [function (require, module, exports) {
-      'use strict';
-      module.exports = function () {
-        return typeof window !== 'undefined' && typeof document !== 'undefined' && typeof document.createElement === 'function';
-      };
-
-    }, {}],
-    6: [function (require, module, exports) {
-      module.exports = extend
-
-      function extend(target) {
-        for (var i = 1; i < arguments.length; i++) {
-          var source = arguments[i]
-
-          for (var key in source) {
-            if (source.hasOwnProperty(key)) {
-              target[key] = source[key]
-            }
-          }
-        }
-
-        return target
-      }
-
-    }, {}]
-  }, {}, [1])(1)
-});
-
+(function(l){"object"===typeof exports&&"undefined"!==typeof module?module.exports=l():"function"===typeof define&&define.amd?define([],l):("undefined"!==typeof window?window:"undefined"!==typeof global?global:"undefined"!==typeof self?self:this).SmartBanner=l()})(function(){return function c(f,g,e){function b(h,d){if(!g[h]){if(!f[h]){var k="function"==typeof require&&require;if(!d&&k)return k(h,!0);if(a)return a(h,!0);k=Error("Cannot find module '"+h+"'");throw k.code="MODULE_NOT_FOUND",k;}k=g[h]=
+{exports:{}};f[h][0].call(k.exports,function(a){var d=f[h][1][a];return b(d?d:a)},k,k.exports,c,f,g,e)}return g[h].exports}for(var a="function"==typeof require&&require,d=0;d<e.length;d++)b(e[d]);return b}({1:[function(c,f,g){var e=c("xtend/mutable"),b=c("component-query"),a=c("get-doc"),d=a&&a.documentElement,h=c("cookie-cutter"),m={ios:{appMeta:"apple-itunes-app-meta",iconRels:["apple-touch-icon-precomposed","apple-touch-icon"],getStoreLink:function(){return this.options.url.ios}},android:{appMeta:"google-play-app",
+iconRels:["android-touch-icon","apple-touch-icon-precomposed","apple-touch-icon"],getStoreLink:function(){return this.options.url.android}}};c=function(a){var b=navigator.userAgent;this.options=e({},{daysHidden:15,daysReminder:90,appStoreLanguage:"us",button:"OPEN",store:{ios:"On the App Store",android:"In Google Play"},price:{ios:"FREE",android:"FREE"},force:!1},a||{});if(this.options.force)this.type=this.options.force;else if(null!==b.match(/iPad|iPhone|iPod/i)){if(null!==b.match(/Safari/i)||null!==
+b.match(/CriOS/i)||6>Number(b.substr(b.indexOf("OS ")+3,3).replace("_",".")))this.type="ios"}else null!==b.match(/Android/i)&&(this.type="android");!this.type||navigator.standalone||h.get("smartbanner-closed")||h.get("smartbanner-installed")||(e(this,m[this.type]),this.parseAppId()&&(this.create(),this.show()))};c.prototype={constructor:c,create:function(){var d=this.getStoreLink(),h=this.options.price[this.type]+" - "+this.options.store[this.type];if(this.options.icon)var c=this.options.icon;else for(var e=
+0;e<this.iconRels.length;e++){var g=b('link[rel="'+this.iconRels[e]+'"]');if(g){c=g.getAttribute("href");break}}var f=a.createElement("div");f.className="smartbanner smartbanner-"+this.type;f.innerHTML='<div class="smartbanner-container"><a href="javascript:void(0);" class="smartbanner-close">&times;</a><span class="smartbanner-icon" style="background-image: url('+c+')"></span><div class="smartbanner-info"><div class="smartbanner-title">'+this.options.title+'</div><div class="smartbanner-description">'+
+this.options.description+"</div><div>"+this.options.author+"</div><span>"+h+'</span></div><a href="'+d+'" class="smartbanner-button"><span class="smartbanner-button-text">'+this.options.button+"</span></a></div>";a.querySelector(".body-wrapper")?(a.querySelector(".body-wrapper").appendChild(f),a.body.classList.add("smart-banner")):a&&a.addEventListener("DOMContentLoaded",function(){a.querySelector(".body-wrapper").appendChild(f);a.body.classList.add("smart-banner")});b(".smartbanner-button",f).addEventListener("click",
+this.install.bind(this),!1);b(".smartbanner-close",f).addEventListener("click",this.close.bind(this),!1)},hide:function(){d.classList.remove("smartbanner-show")},show:function(){d.classList.add("smartbanner-show")},close:function(){this.hide();a.body.classList.remove("smart-banner");h.set("smartbanner-closed","true",{path:"/",expires:+new Date+864E5*this.options.daysHidden})},install:function(){this.hide();h.set("smartbanner-installed","true",{path:"/",expires:+new Date+864E5*this.options.daysReminder})},
+parseAppId:function(){var a=b('meta[name="'+this.appMeta+'"]');if(a)return this.appId="windows"===this.type?a.getAttribute("content"):/app-id=([^\s,]+)/.exec(a.getAttribute("content"))[1]}};f.exports=c},{"component-query":2,"cookie-cutter":3,"get-doc":4,"xtend/mutable":6}],2:[function(c,f,g){function e(b,a){return a.querySelector(b)}g=f.exports=function(b,a){a=a||document;return e(b,a)};g.all=function(b,a){a=a||document;return a.querySelectorAll(b)};g.engine=function(b){if(!b.one)throw Error(".one callback required");
+if(!b.all)throw Error(".all callback required");e=b.one;g.all=b.all;return g}},{}],3:[function(c,f,g){g=f.exports=function(e){e||(e={});"string"===typeof e&&(e={cookie:e});void 0===e.cookie&&(e.cookie="");return{get:function(b){for(var a=e.cookie.split(/;\s*/),d=0;d<a.length;d++){var c=a[d].split("=");if(unescape(c[0])===b)return unescape(c[1])}},set:function(b,a,d){d||(d={});b=escape(b)+"="+escape(a);d.expires&&(b+="; expires="+d.expires);d.path&&(b+="; path="+escape(d.path));return e.cookie=b}}};
+"undefined"!==typeof document&&(c=g(document),g.get=c.get,g.set=c.set)},{}],4:[function(c,f,g){c=c("has-dom");f.exports=c()?document:null},{"has-dom":5}],5:[function(c,f,g){f.exports=function(){return"undefined"!==typeof window&&"undefined"!==typeof document&&"function"===typeof document.createElement}},{}],6:[function(c,f,g){f.exports=function(b){for(var a=1;a<arguments.length;a++){var d=arguments[a],c;for(c in d)e.call(d,c)&&(b[c]=d[c])}return b};var e=Object.prototype.hasOwnProperty},{}]},{},[1])(1)});
